Pf readable (#573)
- wildduck: make pf rules more self-documenting
- dovecot: make pf rules more self-documenting
- mt: install jail.conf file with store_config, so theres a .mt6 laying
nearby
- zonemta: add update support

+++ provision/zonemta.sh
@@ -11,8 +11,12 @@ export JAIL_FSTAB=""
 install_zonemta_webadmin()
 {
 	tell_status "installing ZoneMTA webadmin"
-	stage_exec bash -c "cd /data && git clone https://github.com/zone-eu/zmta-webadmin.git admin"
-	stage_exec bash -c "cd /data/admin && npm install --production"
+	if [ ! -e "$STAGE_MNT/data/admin" ]; then
+		stage_exec bash -c "cd /data && git clone https://github.com/zone-eu/zmta-webadmin.git admin"
+		stage_exec bash -c "cd /data/admin && npm install --production"
+	else
+		stage_exec bash -c "cd /data/admin && git pull && npm install --production"
+	fi
 
 	sed -i '' \
 		-e "/^mongo/ s/127.0.0.1/$(get_jail_ip mongodb)/" \
@@ -27,11 +31,15 @@ install_zonemta()
 	stage_pkg_install npm-node20 git-tiny
 
 	tell_status "installing ZoneMTA"
-	stage_exec bash -c "cd /data && git clone https://github.com/zone-eu/zone-mta-template.git zone-mta"
-	stage_exec bash -c "cd /data/zone-mta && npm install eslint --save-dev"
-	stage_exec bash -c "cd /data/zone-mta && npm init"
-	stage_exec bash -c "cd /data/zone-mta && npm install --production"
-	stage_exec bash -c "cd /data/zone-mta && npm install zonemta-wildduck --save"
+	if [ ! -e "$STAGE_MNT/data/zone-mta" ]; then
+		stage_exec bash -c "cd /data && git clone https://github.com/zone-eu/zone-mta-template.git zone-mta"
+		stage_exec bash -c "cd /data/zone-mta && npm install eslint --save-dev"
+		stage_exec bash -c "cd /data/zone-mta && npm init"
+		stage_exec bash -c "cd /data/zone-mta && npm install --production"
+		stage_exec bash -c "cd /data/zone-mta && npm install zonemta-wildduck --save"
+	else
+		stage_exec bash -c "cd /data/zone-mta && git pull && npm install --production"
+	fi
 
 	sed -i '' \
 		-e "/^mongo/ s/127.0.0.1/$(get_jail_ip mongodb)/" \
@@ -53,7 +61,7 @@ configure_zonemta()
 	stage_exec npm install -g pm2
 	stage_exec pm2 startup
 	stage_sysrc pm2_toor_enable=YES
-	service pm2_toor start
+	stage_exec service pm2_toor start
 
 	tell_status "TODO: configure zonemta-wildduck"
 	echo "https://github.com/nodemailer/zonemta-wildduck"
